[direct] move to package kyo.directInternal
