+++ kyo-direct/shared/src/main/scala/kyo/Direct.scala
@@ -4,6 +4,8 @@ import cps.CpsMonad
 import cps.CpsMonadContext
 import cps.async
 import directInternal.KyoCpsMonad
+import directInternal.Trees
+import directInternal.Validate
 import kyo.Ansi.*
 import kyo.kernel.internal.Safepoint
 import scala.annotation.tailrec
@@ -190,7 +192,7 @@ private def impl[A: Type](body: Expr[A])(using quotes: Quotes): Expr[Any] =
             '{
 
                 given KyoCpsMonad[s] = KyoCpsMonad[s]
-                import asyncShiftInternal.given
+                import kyo.directInternal.asyncShift.given
 
                 async {
                     ${ transformedBody.asExprOf[A] }
@@ -198,25 +200,3 @@ private def impl[A: Type](body: Expr[A])(using quotes: Quotes): Expr[Any] =
             }
     end match
 end impl
-
-object directInternal:
-    given Frame = Frame.internal
-    class KyoCpsMonad[S]
-        extends CpsMonadContext[[A] =>> A < S]
-        with CpsMonad[[A] =>> A < S]
-        with Serializable:
-
-        type Context = KyoCpsMonad[S]
-
-        override def monad: CpsMonad[[A] =>> A < S] = this
-
-        override def apply[A](op: Context => A < S): A < S = op(this)
-
-        override def pure[A](t: A): A < S = t
-
-        override def map[A, B](fa: A < S)(f: A => B): B < S = flatMap(fa)(f)
-
-        override def flatMap[A, B](fa: A < S)(f: A => B < S): B < S = fa.flatMap(f)
-    end KyoCpsMonad
-
-end directInternal
